divide tf and pr pruning test in CI to detect dependency issue
