+++ test/pruning_with_pt/pruning_1.x_v2/test_pruning_regs.py
