+++ test/pruning_with_pt/pruning_2.x/test_auto_excluding_classifier.py
@@ -1,12 +1,6 @@
 import unittest
 
-import torch
-import torchvision
 import torch.nn as nn
-import sys
-sys.path.insert(0, './')
-from neural_compressor.data import Datasets
-from neural_compressor.data.dataloaders.pytorch_dataloader import PyTorchDataLoader
 from transformers import (
     AutoConfig,
     AutoModelForSequenceClassification,
