+++ test/pruning_with_pt/pruning_2.x/test_pruning_config.py
@@ -4,48 +4,37 @@ import torch
 import torchvision
 import torch.nn as nn
 import sys
-sys.path.insert(0, './')
+sys.path.insert(0, '../../pruning_2.x/')
 from neural_compressor.data import Datasets
 from neural_compressor.data.dataloaders.pytorch_dataloader import PyTorchDataLoader
 from neural_compressor import WeightPruningConfig
 from neural_compressor.training import prepare_compression
+from neural_compressor.utils import logger
 
-
-class TestPruningPatterns(unittest.TestCase):
+class TestPytorchPruning(unittest.TestCase):
     model = torchvision.models.resnet18()
 
-    def test_pruning_pattern(self):
+    def test_pruning_class_config(self):
         local_configs = [
             {
-                "op_names": ['layer1.*'],
-                'target_sparsity': 0.5,
-                "pattern": '5:8',
-                "pruning_type": "magnitude"
-            },
-            {
-                "op_names": ['layer2.*'],
-                "pattern": '1xchannel',
-                "pruning_scope": "global"
-            },
-            {
-                "start_step": 2,
-                "end_step": 20,
-                "op_names": ['layer3.*'],
-                'target_sparsity': 0.666666,
-                'pattern': '4x2',
+                "op_names": ['layer1.*', 'layer2.*'],
+                "excluded_op_names": ['downsample.*'],
+                'target_sparsity': 0.6,
+                "pattern": 'channelx1',
                 "pruning_type": "snip_progressive",
-                "pruning_frequency": 5
+                "pruning_scope": "local",
+                "start_step": 0,
+                "end_step": 10
+            },
+            {
+                "op_names": ['layer3.*'],
+                "pruning_type": "pattern_lock"
             }
         ]
         config = WeightPruningConfig(
             local_configs,
+            pruning_frequency=2,
             target_sparsity=0.8,
-            sparsity_decay_type="cos",
-            excluded_op_names=["downsample.*"],
-            pruning_scope="local",
-            min_sparsity_ratio_per_op=0.1,
-            start_step=1,
-            end_step=10
         )
         compression_manager = prepare_compression(model=self.model, confs=config)
         compression_manager.callbacks.on_train_begin()
@@ -53,11 +42,17 @@ class TestPruningPatterns(unittest.TestCase):
         criterion = nn.CrossEntropyLoss()
         optimizer = torch.optim.SGD(self.model.parameters(), lr=0.0001)
         datasets = Datasets('pytorch')
-        dummy_dataset = datasets['dummy'](shape=(10, 3, 224, 224), low=0., high=1., label=True)
+        dummy_dataset = datasets['dummy'](shape=(12, 3, 224, 224), low=0., high=1., label=True)
         dummy_dataloader = PyTorchDataLoader(dummy_dataset)
 
-        compression_manager.callbacks.on_train_begin()
-        for epoch in range(5):
+        logger.info(compression_manager.callbacks.callbacks_list[0].pruners)
+        assert compression_manager.callbacks.callbacks_list[0].pruners[0].config['pruning_frequency'] == 2
+        assert compression_manager.callbacks.callbacks_list[0].pruners[0].config['target_sparsity'] == 0.6
+        assert compression_manager.callbacks.callbacks_list[0].pruners[1].config['target_sparsity'] == 0.8
+        assert compression_manager.callbacks.callbacks_list[0].pruners[0].config['pattern'] == "channelx1"
+        assert compression_manager.callbacks.callbacks_list[0].pruners[1].config['pruning_type'] == 'pattern_lock'
+
+        for epoch in range(1):
             self.model.train()
             compression_manager.callbacks.on_epoch_begin(epoch)
             local_step = 0
