+++ test/pruning_with_pt/pruning_2.x/test_pruning_criteria.py
@@ -4,7 +4,7 @@ import torch
 import torchvision
 import torch.nn as nn
 import sys
-sys.path.insert(0, './')
+sys.path.insert(0, '../../pruning_2.x/')
 from neural_compressor.data import Datasets
 from neural_compressor.data.dataloaders.pytorch_dataloader import PyTorchDataLoader
 from neural_compressor import WeightPruningConfig
