+++ test/pruning_with_pt/pruning_2.x/test_pruning_patterns.py
@@ -4,55 +4,60 @@ import torch
 import torchvision
 import torch.nn as nn
 import sys
-sys.path.insert(0, './')
+sys.path.insert(0, '../../pruning_2.x/')
 from neural_compressor.data import Datasets
 from neural_compressor.data.dataloaders.pytorch_dataloader import PyTorchDataLoader
 from neural_compressor import WeightPruningConfig
 from neural_compressor.training import prepare_compression
 
-local_schedulers_config = [
-    {
-        "start_step": 0,
-        "end_step": 2,
-        "pruning_type": "magnitude",
-        "op_names": ['layer1.*'],
-        "excluded_op_names": ['layer2.*'],
-        "pruning_scope": "global",
-        "target_sparsity": 0.5,
-        "pattern": "4x1"
-    },
-    {
-        "start_step": 1,
-        "end_step": 10,
-        "target_sparsity": 0.5,
-        "pruning_type": "snip_momentum",
-        "pruning_frequency": 2,
-        "op_names": ['layer2.*'],
-        "pruning_scope": "local",
-        "target_sparsity": 0.75,
-        "pattern": "32x1",
-        "sparsity_decay_type": "exp"
-    }
-]
 
-fake_snip_config = WeightPruningConfig(local_schedulers_config, target_sparsity=0.9, start_step=0, \
-                                       end_step=10, pruning_frequency=1, sparsity_decay_type="exp")
-
-
-class TestPruningCriteria(unittest.TestCase):
+class TestPruningPatterns(unittest.TestCase):
     model = torchvision.models.resnet18()
 
-    def test_pruning_schedulers(self):
-
-        compression_manager = prepare_compression(model=self.model, confs=fake_snip_config)
+    def test_pruning_pattern(self):
+        local_configs = [
+            {
+                "op_names": ['layer1.*'],
+                'target_sparsity': 0.5,
+                "pattern": '5:8',
+                "pruning_type": "magnitude"
+            },
+            {
+                "op_names": ['layer2.*'],
+                "pattern": '1xchannel',
+                "pruning_scope": "global"
+            },
+            {
+                "start_step": 2,
+                "end_step": 20,
+                "op_names": ['layer3.*'],
+                'target_sparsity': 0.666666,
+                'pattern': '4x2',
+                "pruning_type": "snip_progressive",
+                "pruning_frequency": 5
+            }
+        ]
+        config = WeightPruningConfig(
+            local_configs,
+            target_sparsity=0.8,
+            sparsity_decay_type="cos",
+            excluded_op_names=["downsample.*"],
+            pruning_scope="local",
+            min_sparsity_ratio_per_op=0.1,
+            start_step=1,
+            end_step=10
+        )
+        compression_manager = prepare_compression(model=self.model, confs=config)
         compression_manager.callbacks.on_train_begin()
+
         criterion = nn.CrossEntropyLoss()
         optimizer = torch.optim.SGD(self.model.parameters(), lr=0.0001)
         datasets = Datasets('pytorch')
         dummy_dataset = datasets['dummy'](shape=(10, 3, 224, 224), low=0., high=1., label=True)
         dummy_dataloader = PyTorchDataLoader(dummy_dataset)
+
         compression_manager.callbacks.on_train_begin()
-        for epoch in range(2):
+        for epoch in range(5):
             self.model.train()
             compression_manager.callbacks.on_epoch_begin(epoch)
             local_step = 0
