+++ test/pruning_with_pt/pruning_2.x/test_pruning_types.py
@@ -4,64 +4,70 @@ import torch
 import torchvision
 import torch.nn as nn
 import sys
-sys.path.insert(0, './')
+sys.path.insert(0, '../../pruning_2.x/')
 from neural_compressor.data import Datasets
 from neural_compressor.data.dataloaders.pytorch_dataloader import PyTorchDataLoader
 from neural_compressor import WeightPruningConfig
 from neural_compressor.training import prepare_compression
 
+local_types_config = [
+    {
+        "start_step": 0,
+        "end_step": 0,
+        "pruning_type": "pattern_lock",
+        "op_names": ['layer1.*'],
+        "excluded_op_names": ['layer2.*'],
+        "pruning_scope": "global"
+    },
+    {
+        "start_step": 1,
+        "end_step": 1,
+        "target_sparsity": 0.5,
+        "pruning_type": "snip_momentum_progressive",
+        "pruning_frequency": 2,
+        "op_names": ['layer2.*'],
+        "pruning_scope": "local",
+        "pattern": "4x1",
+        "sparsity_decay_type": "exp"
+    },
+    {
+        "start_step": 2,
+        "end_step": 8,
+        "target_sparsity": 0.8,
+        "pruning_type": "snip_progressive",
+        "pruning_frequency": 1,
+        "op_names": ['layer3.*'],
+        "pruning_scope": "local",
+        "pattern": "16x1",
+        "sparsity_decay_type": "cube"
+    },
+    {
+        "start_step": 0,
+        "end_step": 0,
+        "pruning_type": "pattern_lock",
+        "op_names": ['layer4.*'],
+        "pattern": "2:4",
+        "pruning_scope": "global"
+    },
+]
 
-class TestPruningPatterns(unittest.TestCase):
+fake_snip_config = WeightPruningConfig(local_types_config, target_sparsity=0.9, start_step=0, \
+                                       end_step=10, pruning_frequency=3, sparsity_decay_type="exp")
+
+
+class TestPruningTypes(unittest.TestCase):
     model = torchvision.models.resnet18()
 
-    def test_pruning_pattern(self):
-        local_configs = [
-            {
-                "op_names": ['layer1.*'],
-                'target_sparsity': 0.75,
-                "pattern": '6:8',
-                "pruning_type": "magnitude_progressive"
-            },
-            {
-                "op_names": ['layer2.*'],
-                "pattern": '1xchannel',
-                "pruning_scope": "global"
-            },
-            {
-                "start_step": 2,
-                "end_step": 20,
-                "op_names": ['layer3.*'],
-                'target_sparsity': 0.666666,
-                'pattern': '4x2',
-                "pruning_type": "snip_progressive",
-                "pruning_frequency": 5
-            }
-        ]
-        config = WeightPruningConfig(
-            local_configs,
-            target_sparsity=0.8,
-            sparsity_decay_type="cos",
-            excluded_op_names=["downsample.*"],
-            pruning_scope="local",
-            min_sparsity_ratio_per_op=0.1,
-            start_step=1,
-            end_step=10
-        )
-        compression_manager = prepare_compression(model=self.model, confs=config)
+    def test_pruning_types(self):
+        compression_manager = prepare_compression(model=self.model, confs=fake_snip_config)
         compression_manager.callbacks.on_train_begin()
-        # fix code coverage
-        compression_manager.callbacks.callbacks_list[0].pruners[-1].progressive_configs['progressive_type'] = "linear"
-        compression_manager.callbacks.callbacks_list[0].pruners[-1].progressive_configs['use_global'] = False
-        compression_manager.callbacks.callbacks_list[0].pruners[-1].progressive_logger = True
-
         criterion = nn.CrossEntropyLoss()
         optimizer = torch.optim.SGD(self.model.parameters(), lr=0.0001)
         datasets = Datasets('pytorch')
         dummy_dataset = datasets['dummy'](shape=(10, 3, 224, 224), low=0., high=1., label=True)
         dummy_dataloader = PyTorchDataLoader(dummy_dataset)
-
         compression_manager.callbacks.on_train_begin()
-        for epoch in range(5):
+        for epoch in range(2):
             self.model.train()
             compression_manager.callbacks.on_epoch_begin(epoch)
             local_step = 0
