+++ test/pruning_with_pt/pruning_2_plus.x/test_pruning_retrain_free.py
@@ -1,22 +1,15 @@
 import unittest
 
-import torch
 import torchvision
 import torch.nn as nn
-import sys
-sys.path.insert(0, './')
 from neural_compressor.data import Datasets
 from neural_compressor.data.dataloaders.pytorch_dataloader import PyTorchDataLoader
 from neural_compressor import WeightPruningConfig
-from transformers import (AutoModelForCausalLM)
 
 
 class TestPruning(unittest.TestCase):
-    # model = torchvision.models.resnet18()
     model = torchvision.models.vit_b_16()
-    # model = AutoModelForCausalLM.from_pretrained(
-    #         "facebook/opt-125m",
-    #         )
+
     def test_pruning_basic(self):
         local_configs = [
             {
@@ -74,6 +67,7 @@ class TestPruning(unittest.TestCase):
                 pruning.on_step_end()
                 local_step += 1
         pruning.on_train_end()
+
 
 if __name__ == "__main__":
     unittest.main()
